+++ spec/fortnox/api/repositories/base_spec.rb
@@ -2,6 +2,8 @@
 
 require 'spec_helper'
 require 'fortnox/api'
+require 'jwt'
+require 'timecop'
 
 describe Fortnox::API::Repository::Base do
   before do
@@ -19,332 +21,451 @@ describe Fortnox::API::Repository::Base do
     Fortnox::API::Registry._container.delete('repositorybasetest')
   end
 
-  let(:access_token) { '3f08d038-f380-4893-94a0-a08f6e60e67a' }
-  let(:access_token2) { '89feajou-sif8-8f8u-29ja-xdfniokeniod' }
-  let(:client_secret) { 'P5K5vE3Kun' }
+  let(:token_store) do
+    Class.new do
+      def access_token
+        # NOTE: Use Timecop with time_when_mocked_jwt_is_valid to use the mocked JWT
+        JWT.encode({ exp: 1_654_159_274 }, nil, 'none')
+      end
+
+      def refresh_token
+        'test-refresh-token'
+      end
+
+      def access_token=(token); end
+
+      def refresh_token=(token); end
+    end
+  end
+  let(:time_when_mocked_jwt_is_valid) { Time.at(1_654_150_000) }
+  let(:time_when_mocked_jwt_has_expired) { Time.at(1_654_159_275) }
+
+  let(:client_id) { 'test-client-id' }
+  let(:client_secret) { 'test-client-secret' }
   let(:repository) { Repository::Test.new }
-  let(:application_json) {}
-  let(:headers) do
-    {
-      'Access-Token' => access_token,
-      'Client-Secret' => client_secret,
-      'Content-Type' => 'application/json',
-      'Accept' => 'application/json'
-    }
+
+  shared_examples_for 'missing configuration' do
+    it 'raises an error' do
+      expect do
+        Repository::Test.new.get('/test')
+      end.to raise_error(Fortnox::API::MissingConfiguration, error_message)
+    end
   end
 
-  describe 'creation' do
-    shared_examples_for 'missing configuration' do
-      subject { -> { Repository::Test.new(token_store: :store1).get('nonsense') } }
+  shared_examples_for 'raises exception' do
+    it 'raises an error' do
+      expect do
+        Repository::Test.new.get('/test')
+      end.to raise_error(Fortnox::API::Exception, error_message)
+    end
+  end
 
-      let(:error) { Fortnox::API::MissingConfiguration }
-
-      it { is_expected.to raise_error(error, /#{message}/) }
+  describe 'configuration' do
+    context 'without any configuration' do
+      include_examples 'missing configuration' do
+        let(:error_message) { /There is no token store named "default"/ }
+      end
     end
 
-    context 'without base url' do
+    context 'without client id' do
       include_examples 'missing configuration' do
         before do
           Fortnox::API.configure do |config|
-            config.base_url = nil
             config.client_secret = client_secret
-            config.access_tokens = { store1: access_token }
+            config.token_stores = { default: token_store.new }
           end
         end
 
-        let(:message) { 'have to provide a base url' }
+        let(:error_message) { /have to provide your client id/ }
       end
     end
 
-    context 'without client secret' do
+    context 'without token stores' do
       include_examples 'missing configuration' do
         before do
           Fortnox::API.configure do |config|
-            config.client_secret = nil
-            config.access_tokens = { store1: access_token }
+            config.client_id = client_id
+            config.client_secret = client_secret
           end
         end
 
-        let(:message) { 'have to provide your client secret' }
+        let(:error_message) { /There is no token store named "default"/ }
       end
     end
-  end
 
-  describe '#next_access_token' do
-    before { Fortnox::API.configure { |conf| conf.client_secret = client_secret } }
-
-    context 'with default token store' do
-      context 'with one access token' do
-        subject { repository.next_access_token }
-
-        before { Fortnox::API.configure { |conf| conf.access_token = access_token } }
-
-        it { is_expected.to eql(access_token) }
-      end
-
-      context 'with one access token and two subsequent requests' do
-        subject { repository.next_access_token }
-
+    context 'without client_secret' do
+      include_examples 'missing configuration' do
         before do
-          Fortnox::API.configure { |conf| conf.access_token = access_token }
-          repository.next_access_token
+          Fortnox::API.configure do |config|
+            config.client_id = client_id
+            config.token_stores = { default: token_store.new }
+          end
         end
 
-        it 'still uses the same token' do
-          is_expected.to eql(access_token)
-        end
-      end
-
-      context 'with multiple access tokens' do
-        subject { access_tokens }
-
-        before { Fortnox::API.configure { |conf| conf.access_tokens = access_tokens } }
-        let(:access_tokens) { [access_token, access_token2] }
-
-        it { is_expected.to include(repository.next_access_token) }
-
-        it 'changes token on next request' do
-          token1 = repository.next_access_token
-          token2 = repository.next_access_token
-
-          expect(token1).not_to eql(token2)
-        end
-
-        it 'circulates tokens' do
-          token1 = repository.next_access_token
-          repository.next_access_token
-          token3 = repository.next_access_token
-
-          expect(token1).to eql(token3)
-        end
+        let(:error_message) { /have to provide your client secret/ }
       end
     end
 
-    context 'with multiple stores' do
-      subject { repository.next_access_token }
-
+    context 'with multiple token stores' do
       before do
         Fortnox::API.configure do |config|
-          config.access_tokens = { store1: access_token, store2: access_token2 }
+          config.client_id = client_id
+          config.client_secret = client_secret
+        end
+
+        stub_request(
+          :get,
+          'https://api.fortnox.se/3/test'
+        ).to_return(
+          status: 200
+        )
+      end
+
+      context 'with default token store' do
+        before do
+          Fortnox::API.configure do |config|
+            config.token_stores = { default: token_store.new, another: token_store.new }
+          end
+          Timecop.freeze(time_when_mocked_jwt_is_valid)
+        end
+
+        after { Timecop.return }
+
+        it 'does not raise an error' do
+          expect { Repository::Test.new.get('/test') }.not_to raise_error
         end
       end
 
-      describe 'first token store' do
-        let(:repository) { Repository::Test.new(token_store: :store1) }
+      context 'without default token store' do
+        before do
+          Fortnox::API.configure do |config|
+            config.token_stores = { another: token_store.new }
+          end
 
-        it { is_expected.to eql access_token }
+          stub_request(
+            :get,
+            'https://api.fortnox.se/3/test'
+          ).to_return(
+            status: 200
+          )
+
+          Timecop.freeze(time_when_mocked_jwt_is_valid)
+        end
+
+        after { Timecop.return }
+
+        it 'raises an error when trying to use default token store' do
+          expect do
+            Repository::Test.new.get('/test')
+          end.to raise_error(Fortnox::API::MissingConfiguration,
+                             /There is no token store named "default"/)
+        end
+
+        it 'does not raise an error when using an explicit token store' do
+          expect do
+            Repository::Test.new(token_store: :another).get('/test')
+          end.not_to raise_error
+        end
       end
 
-      describe 'second token store' do
-        let(:repository) { Repository::Test.new(token_store: :store2) }
+      context 'with invalid token store name' do
+        before do
+          Fortnox::API.configure do |config|
+            config.token_stores = { default: token_store.new, another: token_store.new }
+          end
+        end
 
-        it { is_expected.to eql access_token2 }
+        it 'does not raise an error' do
+          expect do
+            Repository::Test.new(token_store: :invalid).get('/test')
+          end.to raise_error(Fortnox::API::MissingConfiguration,
+                             /There is no token store named "invalid"/)
+        end
       end
     end
   end
 
   describe '#access_tokens' do
-    subject(:access_tokens) { repository.access_tokens }
-
-    before { Fortnox::API.configure { |conf| conf.client_secret = client_secret } }
-
-    let(:token_store_not_present) do
-      "There is no token store named #{token_store.inspect}. Available stores are #{available_stores}."
-    end
-
-    let(:error) { Fortnox::API::MissingConfiguration }
-
-    context 'with non existing token store' do
-      subject { -> { access_tokens } }
-
-      before do
-        Fortnox::API.configure { |conf| conf.access_tokens = { some_store: [access_token] } }
+    before do
+      Fortnox::API.configure do |config|
+        config.client_secret = client_secret
+        config.client_id = client_id
       end
-
-      let(:repository) { Repository::Test.new(token_store: token_store) }
-      let(:token_store) { :non_existing_store }
-      let(:available_stores) { [:some_store] }
-
-      it { is_expected.to raise_error(error, token_store_not_present) }
     end
 
-    context 'with no tokens set' do
-      subject { -> { access_tokens } }
-
-      before { Fortnox::API.configure { |conf| conf.access_tokens = {} } }
-      let(:token_store) { :default }
-      let(:available_stores) { [] }
-
-      it { is_expected.to raise_error(error, token_store_not_present) }
-    end
-
-    context 'with one access token in token store' do
-      before { Fortnox::API.configure { |conf| conf.access_token = access_token } }
-      let(:token_store) { :default }
-
-      it { is_expected.to eql(access_token) }
-    end
-
-    context 'with multiple access tokens' do
+    context 'with nil access and refresh tokens' do
       before do
-        Fortnox::API.configure { |conf| conf.access_tokens = [access_token, access_token2] }
-      end
-      let(:token_store) { :default }
+        Fortnox::API.configure do |config|
+          config.token_stores = {
+            default: Class.new do
+              def access_token; end
 
-      it { is_expected.to eql([access_token, access_token2]) }
-    end
+              def refresh_token; end
 
-    context 'with multiple token stores' do
-      before do
-        Fortnox::API.configure do |conf|
-          conf.access_tokens = { store_a: store_a_tokens, store_b: store_b_token }
+              def access_token=; end
+
+              def refresh_token=; end
+            end.new
+          }
         end
       end
-      let(:store_a_tokens) { %w[token_a1 token_a2] }
-      let(:store_b_token) { 'token_b1' }
 
-      context 'with valid store name' do
-        let(:repository) { Repository::Test.new(token_store: :store_a) }
-
-        it { is_expected.to eql(store_a_tokens) }
-      end
-
-      context 'with non collection' do
-        let(:repository) { Repository::Test.new(token_store: :store_b) }
-
-        it { is_expected.to eql(store_b_token) }
-      end
-
-      context 'with invalid store name' do
-        subject { -> { access_tokens } }
-
-        let(:repository) { Repository::Test.new(token_store: :nonsence_store) }
-
-        it { is_expected.to raise_error(error) }
+      it 'raises an error' do
+        expect do
+          repository.access_token
+        end.to raise_error(Fortnox::API::MissingConfiguration, /Refresh token for store "default" is empty/)
       end
     end
-  end
 
-  describe '#check_access_tokens!' do
-    subject { -> { repository.check_access_tokens!(tokens) } }
+    context 'with empty access and refresh tokens' do
+      before do
+        Fortnox::API.configure do |config|
+          config.token_stores = {
+              default: Class.new do
+              def access_token
+                ''
+              end
 
-    before { Fortnox::API.configure { |conf| conf.client_secret = client_secret } }
-    let(:error) { Fortnox::API::MissingConfiguration }
-    let(:message) { "not provided any access tokens in token store #{token_store.inspect}" }
-    let(:token_store) { :default }
+              def refresh_token
+                ''
+              end
 
-    context 'with nil' do
-      let(:tokens) { nil }
+              def access_token=; end
 
-      it { is_expected.to raise_error(error, /#{message}/) }
+              def refresh_token=; end
+            end.new
+          }
+        end
+      end
+
+      it 'raises an error' do
+        expect do
+          repository.access_token
+        end.to raise_error(Fortnox::API::MissingConfiguration, /Refresh token for store "default" is empty/)
+      end
     end
 
-    context 'with empty array' do
-      let(:tokens) { [] }
+    context 'with nonsense access token' do
+      before do
+        Fortnox::API.configure do |config|
+          config.token_stores = {
+            default: Class.new do
+              def access_token
+                'nonsense'
+              end
 
-      it { is_expected.to raise_error(error, /#{message}/) }
+              def refresh_token; end
+
+              def access_token=; end
+
+              def refresh_token=; end
+            end.new
+          }
+        end
+      end
+
+      it 'raises an error' do
+        expect do
+          repository.access_token
+        end.to raise_error(Fortnox::API::Exception, /Could not decode access token for token store "default"/)
+      end
     end
 
-    context 'with an empty, non default, token store' do
-      before { Fortnox::API.configure { |conf| conf.access_tokens = { token_store => tokens } } }
-      let(:repository) { Repository::Test.new(token_store: token_store) }
-      let(:tokens) { [] }
-      let(:token_store) { :store1 }
+    context 'with nonsense refresh token' do
+      before do
+        Fortnox::API.configure do |config|
+          config.token_stores = {
+            default: Class.new do
+              def access_token; end
 
-      it { is_expected.to raise_error(error, /#{message}/) }
+              def refresh_token
+                'nonsense'
+              end
+
+              def access_token=; end
+
+              def refresh_token=; end
+            end.new
+          }
+        end
+
+        stub_request(:post, 'https://apps.fortnox.se/oauth-v1/token')
+          .to_return(
+            body: '{"error":"invalid_grant","error_description":"Invalid refresh token"}',
+            status: [400, 'Bad Request'],
+            headers: { content_type: 'text/html' } # NOTE: Yes, Fortnox API actually returns html here...
+          )
+      end
+
+      # rubocop:disable RSpec/ExampleLength
+      it 'raises an error' do
+        expect { repository.access_token }.to raise_error(
+          Fortnox::API::Exception,
+          'Unable to renew access token. ' \
+          'Response code: 400. ' \
+          'Response message: Bad Request. ' \
+          'Response body: {"error":"invalid_grant","error_description":"Invalid refresh token"}'
+        )
+      end
+      # rubocop:enable RSpec/ExampleLength
     end
 
-    context 'with valid tokens' do
-      let(:tokens) { %w[12345 abcde] }
+    context 'with no access token available' do
+      context 'with expired refresh token' do
+        before do
+          Fortnox::API.configure do |config|
+            config.token_stores = {
+              default: Class.new do
+                def access_token; end
 
-      it { is_expected.not_to raise_error }
+                def refresh_token
+                  '82f569dcbce8a0b79824e398c16ba6be6c9d8f54'
+                end
+
+                def access_token=; end
+
+                def refresh_token=; end
+              end.new
+            }
+
+            stub_request(:post, 'https://apps.fortnox.se/oauth-v1/token')
+              .to_return(
+                body: '{"error":"invalid_grant","error_description":"Invalid refresh token"}',
+                status: [400, 'Bad Request'],
+                headers: { content_type: 'text/html' } # NOTE: Yes, Fortnox API actually returns html here...
+              )
+          end
+        end
+
+        # rubocop:disable RSpec/ExampleLength
+        it 'raises an error' do
+          expect { repository.access_token }.to raise_error(
+            Fortnox::API::Exception,
+            'Unable to renew access token. ' \
+            'Response code: 400. ' \
+            'Response message: Bad Request. ' \
+            'Response body: {"error":"invalid_grant","error_description":"Invalid refresh token"}'
+          )
+        end
+        # rubocop:enable RSpec/ExampleLength
+      end
+    end
+
+    context 'with invalid access token' do
+      before do
+        Fortnox::API.configure do |config|
+          config.token_stores = {
+            default: Class.new do
+              def access_token
+                'invalid-token'
+              end
+
+              def refresh_token
+                'not-used'
+              end
+
+              def access_token=(token); end
+
+              def refresh_token=(token); end
+            end.new
+          }
+        end
+      end
+
+      it 'raises an error' do
+        expect do
+          repository.access_token
+        end.to raise_error(Fortnox::API::Exception,
+                           'Could not decode access token for token store "default"')
+      end
+    end
+
+    context 'with expired access token and valid refersh token' do
+      before do
+        Fortnox::API.configure do |config|
+          config.token_stores = { default: token_store_spy }
+        end
+
+        stub_request(:post, 'https://apps.fortnox.se/oauth-v1/token')
+          .to_return(
+            body: '{"access_token":"new-access-token","refresh_token":"new-refresh-token"}',
+            status: 200,
+            headers: { content_type: 'application/json' }
+          )
+
+        repository.access_token
+
+        Timecop.freeze(time_when_mocked_jwt_has_expired)
+      end
+
+      after { Timecop.return }
+
+      let(:token_store_spy) do
+        Class.new do
+          attr_reader :received_access_token
+          attr_reader :received_refresh_token
+
+          def initialize; end
+
+          def access_token
+            JWT.encode({ exp: 1_654_159_274 }, nil, 'none')
+          end
+
+          def refresh_token
+            'test-refresh-token'
+          end
+
+          def access_token=(token)
+            @received_access_token = token
+          end
+
+          def refresh_token=(token)
+            @received_refresh_token = token
+          end
+        end.new
+      end
+
+      it 'renews access token and stores it in the token store' do
+        expect(token_store_spy.received_access_token).to eq 'new-access-token'
+      end
+
+      it 'renews refresh token and stores it in the token store' do
+        expect(token_store_spy.received_refresh_token).to eq 'new-refresh-token'
+      end
     end
   end
 
   context 'when making a request including the proper headers' do
+    subject { repository.get('/test', body: '') }
+
     before do
-      Fortnox::API.configure do |conf|
-        conf.client_secret = client_secret
-        conf.access_token = access_token
+      Fortnox::API.configure do |config|
+        config.client_id = client_id
+        config.client_secret = client_secret
+        config.token_stores = { default: token_store.new }
       end
 
       stub_request(
         :get,
         'https://api.fortnox.se/3/test'
-      ).with(
-        headers: headers
       ).to_return(
         status: 200
       )
+
+      Timecop.freeze(time_when_mocked_jwt_is_valid)
     end
 
-    subject { repository.get('/test', body: '') }
+    after { Timecop.return }
 
     it { is_expected.to be_nil }
   end
 
-  describe 'making requests with multiple access tokens' do
+  context 'when receiving an error from remote server' do
     before do
-      Fortnox::API.configure do |conf|
-        conf.client_secret = client_secret
-        conf.access_tokens = [access_token, access_token2]
-      end
-
-      stub_request(
-        :get,
-        'https://api.fortnox.se/3/test'
-      ).with(
-        headers: {
-          'Access-Token' => access_token,
-          'Client-Secret' => client_secret,
-          'Content-Type' => 'application/json',
-          'Accept' => 'application/json'
-        }
-      ).to_return(
-        status: 200,
-        body: '1'
-      )
-
-      stub_request(
-        :get,
-        'https://api.fortnox.se/3/test'
-      ).with(
-        headers: {
-          'Access-Token' => access_token2,
-          'Client-Secret' => client_secret,
-          'Content-Type' => 'application/json',
-          'Accept' => 'application/json'
-        }
-      ).to_return(
-        status: 200,
-        body: '2'
-      )
-    end
-
-    context 'with subsequent requests on same object' do
-      let!(:response1) { repository.get('/test', body: '') }
-      let!(:response2) { repository.get('/test', body: '') }
-      let!(:response3) { repository.get('/test', body: '') }
-
-      # rubocop:disable RSpec/MultipleExpectations
-      # All these checks must be run in same it-statement because
-      # of the random starting index.
-      it 'works' do
-        expect(response1).not_to eq(response2)
-        expect(response1).to eq(response3)
-        expect(response3).not_to eq(response2)
-      end
-      # rubocop:enable RSpec/MultipleExpectations
-    end
-  end
-
-  context 'when raising error from remote server' do
-    subject { -> { repository.post('/test', body: '') } }
-
-    before do
-      Fortnox::API.configure do |conf|
-        conf.client_secret = client_secret
-        conf.access_tokens = [access_token, access_token2]
+      Fortnox::API.configure do |config|
+        config.client_id = client_id
+        config.client_secret = client_secret
+        config.token_stores = { default: token_store.new }
       end
 
       stub_request(
@@ -355,20 +476,22 @@ describe Fortnox::API::Repository::Base do
         body: {
           'ErrorInformation' => {
             'error' => 1,
-            'message' => error_message
+            'message' => 'some-error-message'
           }
         }.to_json,
         headers: { 'Content-Type' => 'application/json' }
       )
+
+      Timecop.freeze(time_when_mocked_jwt_is_valid)
     end
 
-    let(:error_message) do
-      'Räkenskapsår finns inte upplagt. '\
-      'För att kunna skapa en faktura krävs det att det finns ett räkenskapsår'
-    end
+    after { Timecop.return }
 
-    it { is_expected.to raise_error(Fortnox::API::RemoteServerError) }
-    it { is_expected.to raise_error(error_message) }
+    it 'raises an error' do
+      expect do
+        repository.post('/test', body: '')
+      end.to raise_error(Fortnox::API::RemoteServerError, 'some-error-message')
+    end
 
     context 'with debugging enabled' do
       around do |example|
@@ -377,7 +500,44 @@ describe Fortnox::API::Repository::Base do
         Fortnox::API.config.debugging = false
       end
 
-      it { is_expected.to raise_error(/\<HTTParty\:\:Request\:0x/) }
+      it do
+        expect do
+          repository.post('/test', body: '')
+        end.to raise_error(/\<HTTParty\:\:Request\:0x/)
+      end
+    end
+  end
+
+  context 'when receiving HTML from remote server' do
+    before do
+      Fortnox::API.configure do |config|
+        config.client_id = client_id
+        config.client_secret = client_secret
+        config.token_stores = { default: token_store.new }
+      end
+
+      stub_request(
+        :get,
+        'https://api.fortnox.se/3nonsense'
+      ).to_return(
+        status: 404,
+        body: "<html>\r\n" \
+              "<head><title>404 Not Found</title></head>\r\n" \
+              "<body>\r\n<center><h1>404 Not Found</h1></center>\r\n" \
+              "<hr><center>nginx</center>\r\n</body>\r\n</html>\r\n",
+        headers: { 'Content-Type' => 'text/html' }
+      )
+
+      Timecop.freeze(time_when_mocked_jwt_is_valid)
+    end
+
+    after { Timecop.return }
+
+    it 'raises an error' do
+      expect do
+        repository.get('nonsense')
+      end.to raise_error(Fortnox::API::RemoteServerError,
+                         %r{Fortnox API's response has content type "text/html"})
     end
   end
 end