+++ spec/fortnox/api/types/housework_types_spec.rb
@@ -10,6 +10,7 @@ require 'fortnox/api/types/order_row'
 # rubocop:disable RSpec/DescribeClass
 describe 'HouseworkTypes', integration: true do
   include Helpers::Configuration
+  include Helpers::Repositories
 
   before { set_api_test_configuration }
 
@@ -28,7 +29,7 @@ describe 'HouseworkTypes', integration: true do
         order_rows: [
           Fortnox::API::Types::OrderRow.new(
             ordered_quantity: 1,
-            article_number: '0000',
+            article_number: '101',
             housework_type: Fortnox::API::Types::HouseworkTypes[type],
             housework: housework
           )
@@ -95,35 +96,6 @@ describe 'HouseworkTypes', integration: true do
   it_behaves_like 'housework type', 'COOKING', TYPE_RUT, legacy: true
   it_behaves_like 'housework type', 'TUTORING', TYPE_RUT, legacy: true
 
-  describe 'without tax reduction type' do
-    subject do
-      cassette = 'orders/housework_without_tax_reduction_type'
-      -> { VCR.use_cassette(cassette) { repository.save(document) } }
-    end
-
-    let(:document) do
-      Fortnox::API::Model::Order.new(
-        customer_number: '1',
-        order_rows: [
-          Fortnox::API::Types::OrderRow.new(
-            ordered_quantity: 1,
-            article_number: '0000',
-            housework_type: Fortnox::API::Types::HouseworkTypes['CONSTRUCTION'],
-            housework: true
-          )
-        ]
-      )
-    end
-
-    let(:error_message) do
-      'Dokument utan angiven skattereduktionstyp får inte innehålla artikelrader med husarbetestyp.'
-    end
-
-    it 'raises an error' do
-      is_expected.to raise_error(Fortnox::API::RemoteServerError, error_message)
-    end
-  end
-
   describe 'with OTHERCOSTS' do
     subject do
       cassette = 'orders/housework_othercoses_invalid'
@@ -137,7 +109,7 @@ describe 'HouseworkTypes', integration: true do
         order_rows: [
           Fortnox::API::Types::OrderRow.new(
             ordered_quantity: 1,
-            article_number: '0000',
+            article_number: '101',
             housework_type: Fortnox::API::Types::HouseworkTypes['OTHERCOSTS'],
             housework: true
           )
@@ -166,7 +138,7 @@ describe 'HouseworkTypes', integration: true do
         order_rows: [
           Fortnox::API::Types::OrderRow.new(
             ordered_quantity: 1,
-            article_number: '0000',
+            article_number: '101',
             housework_type: Fortnox::API::Types::HouseworkTypes[type],
             housework: true
           )