+++ spec/spec_helper.rb
@@ -20,7 +20,6 @@ RSpec.configure do |config|
   config.extend Helpers # Allow access to helpers in describe and context blocks
   config.include Helpers # Allow access to helpers in it and let blocks
 
-  config.include Helpers::Repositories, integration: true
   config.include Matchers::Type, type: :type
 
   config.order = 'random'